kodo: add layer to allow decoded-aware encoding
This patch adds a generator layer that enables the encoder to exploit
status information from the decoder to only include innovative symbols
in encoded packets.

The layer uses the encoder_status_vector layer to check if status
information from the decoder has been added, and if so, calls the new
generate_status_vector_aware() function in the uniform_generator layer.

Signed-off-by: Martin Hundebøll <m.hundeboll@gmail.com>
Signed-off-to: Steinwurf ApS

NB: When only one new packet is added to an encoder that uses binary
field size, there is a 50% risk that the next non-systematic packet
will be the null-vector, so use with care.