kodo: make systematic encoder storage aware
If a user chooses to add symbols to an on-the-fly encoder in a
non-chronological order, the systematic encoder will fail, as it uses an
incremental counter to select the next symbol to encode (systematically).

This patch adds a storage tracker layer that maintains a bitmap of
available symbols, and a systematic encoder layer that uses the former
to choose the next symbol to encode.

Signed-off-by: Martin Hundebøll <m.hundeboll@gmail.com>
Signed-off-to: SteinWurf ApS <contact@steinwurf.com>