kodo: use is_symbol_decoded() in systematic encoding
During the decoding process, there is a difference on pivot symbols and
decoded symbols. This difference is unclear when using the symbol
storage as a fake decoder in the encoder, since the symbol storage has
always decoded symbols and not "just" pivot symbols.

This patch renames the use of is_symbol_pivot() in the
systematic-related layers to is_symbol_decoded(). The difference will
become significant when implementing systematic encoding in a recoder
stack.

Signed-off-by: Martin Hundebøll <m.hundeboll@gmail.com>
Signed-off-to: Steinwurf ApS <contact@steinwurf.com>