kodo: rename feedback related remote_rank() to remote_decoder_rank()
If one uses a systematic recoder with feedback for sliding window
coding, two remote ranks appears: one from the encoder sending symbols
to the recoder, and one from the decoder sending feedback to the
recoder. In this case, the recoder should forward the rank received
from the encoder instead of the rank received from the decoder.

This patch changes the name of the feedback-given remote rank from
remote_rank() to remote_decoder_rank() to allow the
payload_rank_recoder layer to use the correct remote rank.

Signed-off-by: Martin Hundebøll <m.hundeboll@gmail.com>
Signed-off-to: Steinwurf ApS <contact@steinwurf.com>