kodo: use systematic recoder for sliding window codes
With support for partial decoding, the recoder should be able to send
systematic packets when it has decoded symbols available. Also, it
should be able to make use of feedback from the decoder.

This patch changes the recoding stack in rlnc/sliding_window_decoder.hpp
to use the sliding_window_encoder. To allow this, the
storage_aware_systematic_phase layer is changed to use symbols_decoded()
instead of simply rank() when deciding wether to use systematic packets
for not, as the rank() might include seen-but-not-yet-decoded symbols.

Signed-off-by: Martin Hundebøll <m.hundeboll@gmail.com>
Signed-off-to: Steinwurf ApS <contact@steinwurf.com>